Trocando os valores de float para double

+++ src/main/java/br/gov/es/infoplan/dto/strategicProject/StrategicProjectAccumulatedInvestmentDto.java
@@ -3,17 +3,17 @@ package br.gov.es.infoplan.dto.strategicProject;
 public class StrategicProjectAccumulatedInvestmentDto {
 
     private int anoMes;
-    private float custoPrevisto;
-    private float custoPrevistoAcumulado;
-    private float custoRealizado;
-    private float custoRealizadoAcumulado;
+    private double custoPrevisto;
+    private double custoPrevistoAcumulado;
+    private double custoRealizado;
+    private double custoRealizadoAcumulado;
 
     
     public StrategicProjectAccumulatedInvestmentDto() {
     }
 
-    public StrategicProjectAccumulatedInvestmentDto(int anoMes, float custoPrevisto, float custoPrevistoAcumulado,
-            float custoRealizado, float custoRealizadoAcumulado) {
+    public StrategicProjectAccumulatedInvestmentDto(int anoMes, double custoPrevisto, double custoPrevistoAcumulado,
+            double custoRealizado, double custoRealizadoAcumulado) {
         this.anoMes = anoMes;
         this.custoPrevisto = custoPrevisto;
         this.custoPrevistoAcumulado = custoPrevistoAcumulado;
@@ -27,28 +27,28 @@ public class StrategicProjectAccumulatedInvestmentDto {
     public void setAnoMes(int anoMes) {
         this.anoMes = anoMes;
     }
-    public float getCustoPrevisto() {
+    public double getCustoPrevisto() {
         return custoPrevisto;
     }
-    public void setCustoPrevisto(float custoPrevisto) {
+    public void setCustoPrevisto(double custoPrevisto) {
         this.custoPrevisto = custoPrevisto;
     }
-    public float getCustoPrevistoAcumulado() {
+    public double getCustoPrevistoAcumulado() {
         return custoPrevistoAcumulado;
     }
-    public void setCustoPrevistoAcumulado(float custoPrevistoAcumulado) {
+    public void setCustoPrevistoAcumulado(double custoPrevistoAcumulado) {
         this.custoPrevistoAcumulado = custoPrevistoAcumulado;
     }
-    public float getCustoRealizado() {
+    public double getCustoRealizado() {
         return custoRealizado;
     }
-    public void setCustoRealizado(float custoRealizado) {
+    public void setCustoRealizado(double custoRealizado) {
         this.custoRealizado = custoRealizado;
     }
-    public float getCustoRealizadoAcumulado() {
+    public double getCustoRealizadoAcumulado() {
         return custoRealizadoAcumulado;
     }
-    public void setCustoRealizadoAcumulado(float custoRealizadoAcumulado) {
+    public void setCustoRealizadoAcumulado(double custoRealizadoAcumulado) {
         this.custoRealizadoAcumulado = custoRealizadoAcumulado;
     }
